+++ linux-qt/src/TerminalPane.h
@@ -1,9 +1,15 @@
 #pragma once
 
+#include <QFont>
+#include <QWidget>
+
+#ifndef TFX_HAVE_QTERMWIDGET
 #include <QLineEdit>
 #include <QPlainTextEdit>
 #include <QProcess>
-#include <QWidget>
+#endif
+
+class QTermWidget;
 
 class TerminalPane : public QWidget
 {
@@ -12,16 +18,27 @@ class TerminalPane : public QWidget
 public:
     explicit TerminalPane(QWidget *parent = nullptr);
     void setWorkingDirectory(const QString &path);
+    void setContentFont(const QFont &font);
+    void setColorScheme(const QString &name);
 
 signals:
     void closeRequested();
 
-private slots:
-    void runCommand();
+protected:
+    void showEvent(QShowEvent *event) override;
 
 private:
-    QPlainTextEdit *m_output;
-    QLineEdit *m_command;
-    QProcess *m_process;
     QString m_workingDirectory;
+
+#ifdef TFX_HAVE_QTERMWIDGET
+    void startTerminal();
+    QTermWidget *m_term = nullptr;
+    QFont m_font;
+    bool m_started = false;
+#else
+    void runCommand();
+    QPlainTextEdit *m_output = nullptr;
+    QLineEdit *m_command = nullptr;
+    QProcess *m_process = nullptr;
+#endif
 };